requirements.txt
# Employee Monitoring System - Dependencies
# Core dependencies
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0

# Face detection and recognition
face-recognition==1.3.0
dlib==19.24.2

# Database
sqlite3  # Built-in with Python

# Notifications and alerts
plyer==2.1.0
winsound  # Built-in with Python (Windows only)

# Data handling
pandas==2.1.3

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Additional utilities
python-dateutil==2.8.2
requests==2.31.0
werkzeug==3.0.1
